containers/boot/requirements.txt
boto3==1.37.20 
load-dotenv==0.1.0
qdrant-client==1.12.1
requests==2.32.3